refactor styles in different files
